+++ src/px-script/index.js
@@ -4,9 +4,11 @@ import actionList from "./action-list";
 import alert from "./alert";
 import button from "./button";
 import datepicker from "./datepicker";
+import dialog from "./dialog";
 import loader from "./loader";
 import modal from "./modal";
 import rangeslider from "./rangeslider";
+import sheet from "./sheet";
 import script from "./script";
 import tabs from "./tabs";
 import toast from "./toast";
@@ -21,9 +23,11 @@ const px = {
     alert,
     button,
     datepicker,
+    dialog,
     loader,
     modal,
     rangeslider,
+    sheet,
     script,
     tabs,
     toast,
@@ -40,4 +44,4 @@ if (!window.stopPx) {
 }
 
 export default px;
-export { actionList, alert, button, datepicker, loader, modal, rangeslider, script, tabs, toast, topbar, validation };
+export { actionList, alert, button, datepicker, dialog, loader, modal, rangeslider, sheet, script, tabs, toast, topbar, validation };